refactor: reorganize frontend structure
