+++ codespace/frontend/src/assets/images/architecture.png
