+++ codespace/frontend/src/pages/LoginPage.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
-import '../LoginPage.css';
+import '../styles/LoginPage.css';
 
 function LoginPage() {
   const [identifier, setIdentifier] = useState('');
