+++ codespace/frontend/src/pages/SignupPage.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
-import '../SignupPage.css';
+import '../styles/SignupPage.css';
 
 function SignupPage() {
   const [username, setUsername] = useState('');
