+++ codespace/frontend/src/pages/WelcomePage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import '../WelcomePage.css';
+import '../styles/WelcomePage.css';
 
 function WelcomePage() {
   const navigate = useNavigate();
