+++ codespace/frontend/src/styles/Samples.css
